// content/games.generated.ts
export type Category = {
  slug: string;
  label: { es: string; en: string };
  count: number;
};

export const categories: Category[] = [
  {
    "slug": "novedades",
    "label": {
      "es": "Novedades",
      "en": "New Arrivals"
    },
    "count": 10
  },
  {
    "slug": "aventura",
    "label": {
      "es": "Aventura",
      "en": "Adventure"
    },
    "count": 22
  },
  {
    "slug": "casual",
    "label": {
      "es": "Casual",
      "en": "Casual"
    },
    "count": 20
  },
  {
    "slug": "adrenalina",
    "label": {
      "es": "Adrenalina",
      "en": "Adrenaline Rush"
    },
    "count": 14
  },
  {
    "slug": "arcade",
    "label": {
      "es": "Arcade",
      "en": "Arcade"
    },
    "count": 14
  }
];